requirements.txt
tensorflow
matplotlib
fastapi
uvicorn
mangum
pillow
python-multipart
numpy==1.23.1